refactor(validate)!: rename validate_enhanced -> validate_fixes
Clarify the responsibility split between `validate.py` (visual checks
like overflow, overlap, tick crowding, pie label offset) and the
auto-fix companion. The old "enhanced" name was vague — `validate_fixes`
makes it obvious that the submodule provides the auto-fix layer
(`get_fix_suggestions`, `validate_with_fixes`,
`check_agent_requirements`, `generate_validation_report`).

The public function `dm.validate_with_fixes` is unchanged. Direct
submodule access (`dm.validate_enhanced` /
`from dartwork_mpl.validate_enhanced import ...`) now raises
`AttributeError` / `ModuleNotFoundError` — use `dm.validate_fixes`
instead. No backward-compat alias is provided, in line with the 0.4.x
deprecated-surface purge (#87).
