+++ docs/api/validate.rst
@@ -67,10 +67,10 @@ Core Validation
 Enhanced Validation with Auto-Fix
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-The ``validate_enhanced`` module provides advanced validation with automatic fix suggestions,
+The ``validate_fixes`` module provides advanced validation with automatic fix suggestions,
 particularly useful for AI agents and automated pipelines.
 
-.. automodule:: dartwork_mpl.validate_enhanced
+.. automodule:: dartwork_mpl.validate_fixes
    :members: validate_with_fixes, get_fix_suggestions, check_agent_requirements, generate_validation_report
    :noindex:
 
@@ -80,7 +80,7 @@ Example with Auto-Fix
 .. code-block:: python
 
    import dartwork_mpl as dm
-   from dartwork_mpl.validate_enhanced import validate_with_fixes
+   from dartwork_mpl.validate_fixes import validate_with_fixes
 
    # Validate and get fix suggestions
    issues, fixes = validate_with_fixes(fig)
